void* casting;64-bit issue fix
git-svn-id: https://uhpc.nbgisp.com/svn/runnemede/trunk@1536 b8300b04-83cd-4eb7-bcdf-78d2b5523b00

+++ apps/kernels/bfs_smp_for_afl/bfs_smp_step_by_step.c
@@ -52,7 +52,7 @@ int main(int argc, char** argv)
   for(int nq=0; nq < num_queries; ++nq)
     {
       printf("Start vertex = %d\n", start_v);
-      for(int i =0; i< num_explorers; ++i)
+      for(int64_t i =0; i< num_explorers; ++i)
 	{
 	  pthread_create(&all_threads[i], 0, explorer_thread, (void*)i);
 	}
